chore(storage): refactor category structure (#11957)
* chore: refactor storage category structure

* chore: code cleanup

---------
