+++ packages/storage/src/providers/s3/apis/uploadData/multipart/getDataChunker.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { UploadSource } from '../../../../../types';
+import { StorageUploadSourceOptions } from '../../../../../types';
 import {
 	StorageValidationErrorCode,
 	validationErrorMap,
@@ -15,7 +15,10 @@ export type PartToUpload = {
 	size: number;
 };
 
-export const getDataChunker = (data: UploadSource, totalSize?: number) => {
+export const getDataChunker = (
+	data: StorageUploadSourceOptions,
+	totalSize?: number
+) => {
 	const partSize = calculatePartSize(totalSize);
 
 	if (data instanceof Blob) {
